Clean up notions of definitional "strength"
between unforegablility and EUFCMAVES. I previously claimed that
EUFCMAVES was stronger than unforgeability. This is true in the less
obvious sense of being a stronger security claim i.e. CDH is stronger
assumption than discrete log. But is not stronger in the sense that
breaking VES is harder than breaking EUFCMAVES.